Refactor data types and indexing: change BookEntryId and ID fields to uint16 for better memory efficiency, implement new title indexing logic, and add benchmark tests for memory layout optimization.

+++ internal/model/bookentry.sql.go
@@ -22,7 +22,7 @@ FROM books
 `
 
 type BookEntryRow struct {
-	ID         int64
+	ID         uint16
 	Title      string
 	Authors    string
 	AddedAt    time.Time
